+++ packages/block-library/src/query/view.js
@@ -19,46 +19,22 @@ const isValidEvent = ( event ) =>
 	! event.defaultPrevented;
 
 store( 'core/query', {
-	state: {
-		get startAnimation() {
-			return getContext().animation === 'start';
-		},
-		get finishAnimation() {
-			return getContext().animation === 'finish';
-		},
-	},
 	actions: {
 		*navigate( event ) {
 			const ctx = getContext();
 			const { ref } = getElement();
-			const { queryRef } = ctx;
+			const queryRef = ref.closest(
+				'.wp-block-query[data-wp-router-region]'
+			);
 			const isDisabled = queryRef?.dataset.wpNavigationDisabled;
 
 			if ( isValidLink( ref ) && isValidEvent( event ) && ! isDisabled ) {
 				event.preventDefault();
 
-				// Don't announce the navigation immediately, wait 400 ms.
-				const timeout = setTimeout( () => {
-					ctx.message = ctx.loadingText;
-					ctx.animation = 'start';
-				}, 400 );
-
 				const { actions } = yield import(
 					'@wordpress/interactivity-router'
 				);
 				yield actions.navigate( ref.href );
-
-				// Dismiss loading message if it hasn't been added yet.
-				clearTimeout( timeout );
-
-				// Announce that the page has been loaded. If the message is the
-				// same, we use a no-break space similar to the @wordpress/a11y
-				// package: https://github.com/WordPress/gutenberg/blob/c395242b8e6ee20f8b06c199e4fc2920d7018af1/packages/a11y/src/filter-message.js#L20-L26
-				ctx.message =
-					ctx.loadedText +
-					( ctx.message === ctx.loadedText ? '\u00A0' : '' );
-
-				ctx.animation = 'finish';
 				ctx.url = ref.href;
 
 				// Focus the first anchor of the Query block.
@@ -67,8 +43,10 @@ store( 'core/query', {
 			}
 		},
 		*prefetch() {
-			const { queryRef } = getContext();
 			const { ref } = getElement();
+			const queryRef = ref.closest(
+				'.wp-block-query[data-wp-router-region]'
+			);
 			const isDisabled = queryRef?.dataset.wpNavigationDisabled;
 			if ( isValidLink( ref ) && ! isDisabled ) {
 				const { actions } = yield import(
@@ -88,11 +66,6 @@ store( 'core/query', {
 				);
 				yield actions.prefetch( ref.href );
 			}
-		},
-		setQueryRef() {
-			const ctx = getContext();
-			const { ref } = getElement();
-			ctx.queryRef = ref;
 		},
 	},
 } );