+++ phpunit/tests/fonts/font-library/fontLibraryHooks.php
@@ -4,9 +4,12 @@
  *
  * @package WordPress
  * @subpackage Font Library
+ *
+ * @group fonts
+ * @group font-library
  */
+class Tests_Fonts_FontLibraryHooks extends WP_UnitTestCase {
 
-class Tests_Font_Library_Hooks extends WP_UnitTestCase {
 	public function test_deleting_font_family_deletes_child_font_faces() {
 		$font_family_id       = self::factory()->post->create(
 			array(
@@ -33,8 +36,8 @@ class Tests_Font_Library_Hooks extends WP_UnitTestCase {
 
 		wp_delete_post( $font_family_id, true );
 
-		$this->assertNull( get_post( $font_face_id ) );
-		$this->assertNotNull( get_post( $other_font_face_id ) );
+		$this->assertNull( get_post( $font_face_id ), 'Font face post should also have been deleted.' );
+		$this->assertNotNull( get_post( $other_font_face_id ), 'The other post should exist.' );
 	}
 
 	public function test_deleting_font_faces_deletes_associated_font_files() {
@@ -43,8 +46,8 @@ class Tests_Font_Library_Hooks extends WP_UnitTestCase {
 
 		wp_delete_post( $font_face_id, true );
 
-		$this->assertFalse( file_exists( $font_path ) );
-		$this->assertTrue( file_exists( $other_font_path ) );
+		$this->assertFalse( file_exists( $font_path ), 'The font file should have been deleted when the post was deleted.' );
+		$this->assertTrue( file_exists( $other_font_path ), 'The other font file should exist.' );
 	}
 
 	protected function create_font_face_with_file( $filename ) {
@@ -70,7 +73,7 @@ class Tests_Font_Library_Hooks extends WP_UnitTestCase {
 		// @core-merge Use `DIR_TESTDATA` instead of `GUTENBERG_DIR_TESTDATA`.
 		$font_file_path = GUTENBERG_DIR_TESTDATA . 'fonts/' . $font_filename;
 
-		add_filter( 'upload_mimes', array( 'WP_Font_Library', 'set_allowed_mime_types' ) );
+		add_filter( 'upload_mimes', array( 'WP_Font_Utils', 'get_allowed_font_mime_types' ) );
 		add_filter( 'upload_dir', 'wp_get_font_dir' );
 		$font_file = wp_upload_bits(
 			$font_filename,
@@ -78,7 +81,7 @@ class Tests_Font_Library_Hooks extends WP_UnitTestCase {
 			file_get_contents( $font_file_path )
 		);
 		remove_filter( 'upload_dir', 'wp_get_font_dir' );
-		remove_filter( 'upload_mimes', array( 'WP_Font_Library', 'set_allowed_mime_types' ) );
+		remove_filter( 'upload_mimes', array( 'WP_Font_Utils', 'get_allowed_font_mime_types' ) );
 
 		return $font_file;
 	}